+++ docs/Changelog.md
@@ -1,5 +1,7 @@
 # Release Log
 
+- 2025-12-15: Released musicdl v2.7.2 — added support for jamendo and make some improvements.
+
 - 2025-12-11: Released musicdl v2.7.1 — added support for two new sites and fixed several potential bugs.
 
 - 2025-12-10: Released musicdl v2.7.0 — the code has been further refactored, with a large amount of redundant code removed or merged; all supported sites can now download lossless music (for some sites, you need to set your membership cookies in the command line or in the code), the search speed has been greatly optimized, and several problematic sites have been fixed.